% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getScore}
\alias{getScore}
\title{Get differential score for each observation}
\usage{
getScore(object, type = "weight")
}
\arguments{
\item{object}{Object of class inheriting from "mpk".}

\item{type}{String indicating whether the score is associate to a difference in mixture weights
\code{type = "weight"}, or a difference in the mean parameters \code{type = "shift"}.}
}
\value{
A vector with a score for each observation.
}
\description{
Get differential score for each observation
}
\examples{
n = c(250, 250)
p = 4

Y1 = rbind( matrix( rnorm( n[1]*p), ncol = p), matrix( rnorm(n[2]*p) + 3, ncol = p))
Y2 = rbind( matrix( rnorm( n[1]*p), ncol = p), matrix( rnorm(n[2]*p) + 4, ncol = p))
Y = rbind(Y1, Y2)
C = c( rep(1,sum(n)), rep(2,sum(n)))

ans = mpk(Y, C)
score = getScore(ans, type = "shift")
hist(score)
}

