% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotDiff}
\alias{plotDiff}
\title{Plot differences across mixture models}
\usage{
plotDiff(object, type = "weight", dim = c(1, 2),
  group = 1:length(unique(object$data$C)), xlim = range(object$data$Y[,
  dim[1]]), ylim = range(object$data$Y[, dim[2]]), ...)
}
\arguments{
\item{object}{Object of class inheriting from "mpk".}

\item{type}{String indicating whether the data-points are colored according to difference in mixture weights
\code{type = "weight"}, or difference in the mean parameters \code{type = "shift"}.}

\item{dim}{Vector indicating along which dimension the data are projected.
Default is \code{dim = c(1,2)}.}

\item{group}{Vector indicating which groups are plotted. By default all groups are plotted.}

\item{xlim}{the x limits (x1, x2) of the plot.}

\item{xlim}{the y limits of the plot.}
}
\value{
A plot.
}
\description{
Plot differences across mixture models
}
\examples{
n = c(250, 250)
p = 4

Y1 = rbind( matrix( rnorm( n[1]*p), ncol = p), matrix( rnorm(n[2]*p) + 3, ncol = p))
Y2 = rbind( matrix( rnorm( n[1]*p), ncol = p), matrix( rnorm(n[2]*p) + 4, ncol = p))
Y = rbind(Y1, Y2)
C = c( rep(1,sum(n)), rep(2,sum(n)))

ans = mpk(Y, C)
plotDiff(ans, type = "weight")
plotDiff(ans, type = "shift")
}

